% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rides.r
\docType{data}
\name{daily_totals}
\alias{daily_totals}
\title{A year of simulated bicycle ride mileages, aggregated by day}
\format{
A data frame with 178 rows and 2 variables:
\describe{
\item{ride_date}{date the ride occurred}
\item{total_length}{the total length in miles for each day}
}
}
\usage{
daily_totals
}
\description{
Simulated dates and distances of rides occurring in 2009. This is an
aggregation of the \code{\link{rides}} dataset by day.
}
\details{
The dataset contains a total of 3,419 miles spread across 178
unique days. The Eddington number for the year was 29.
}
\seealso{
\link{rides}
}
\keyword{datasets}
