% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durfee.R
\name{durfee}
\alias{durfee}
\title{Compute the side length of a Durfee square}
\usage{
durfee(is)
}
\arguments{
\item{is}{An integer vector representing an integer partition.}
}
\value{
The side length of the Durfee square for that partition.
}
\description{
Compute the side length of a Durfee square
}
