% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_cum}
\alias{E_cum}
\title{Calculate the cumulative Eddington number}
\usage{
E_cum(rides)
}
\arguments{
\item{rides}{A vector of mileage, where each element represents a single
day.}
}
\value{
An integer vector the same length as \code{rides}.
}
\description{
This function is much like \code{\link{E_num}} except it provides
a cumulative Eddington number over the vector rather than a single summary
number.
}
\seealso{
\code{\link{E_next}}, \code{\link{E_num}}, \code{\link{E_req}},
  \code{\link{E_sat}}
}
