% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{E_next}
\alias{E_next}
\title{Get the number of rides required to increment to the next Eddington number}
\usage{
E_next(rides)
}
\arguments{
\item{rides}{A vector of mileage, where each element represents a single
day.}
}
\value{
A named list with the current Eddington number (\code{E}) and the
  number of rides required to increment by one (\code{req}).
}
\description{
Get the number of rides required to increment to the next Eddington number.
}
\seealso{
\code{\link{E_cum}}, \code{\link{E_num}}, \code{\link{E_req}},
  \code{\link{E_sat}}
}
