% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddington.r
\name{E_req}
\alias{E_req}
\title{Determine the number of additional rides required to achieve a specified
Eddington number}
\usage{
E_req(rides, candidate)
}
\arguments{
\item{rides}{A vector of mileage, where each element represents a single
day.}

\item{candidate}{The Eddington number to test for.}
}
\value{
An integer vector of length 1. Returns \code{0L} if \emph{E} is
  already achieved.
}
\description{
Determine the number of additional rides required to achieve a specified
Eddington number.
}
\seealso{
\code{\link{E_cum}}, \code{\link{E_next}}, \code{\link{E_num}},
  \code{\link{E_sat}}
}
