% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddington.r
\name{E_sat}
\alias{E_sat}
\title{Determine if a dataset satisfies a specified Eddington number}
\usage{
E_sat(rides, candidate)
}
\arguments{
\item{rides}{A vector of mileage, where each element represents a single
day.}

\item{candidate}{The Eddington number to test for.}
}
\value{
A logical vector of length 1.
}
\description{
Indicates whether a certain Eddington number is satisfied, given the data.
}
\seealso{
\code{\link{E_cum}}, \code{\link{E_next}}, \code{\link{E_num}},
  \code{\link{E_req}}
}
