% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliometrics.R
\name{h_index}
\alias{h_index}
\alias{bibliometrics}
\alias{i10_index}
\alias{g_index}
\title{Compute several bibliometric indices}
\usage{
h_index(citations, na.rm = FALSE)

i10_index(citations, na.rm = FALSE)

g_index(citations, na.rm = FALSE, is_sorted = FALSE)
}
\arguments{
\item{citations}{A vector of citation counts.}

\item{na.rm}{If \code{TRUE}, \code{NA} values will be filtered out. Otherwise, any \code{NA}
value found in the vector will propagate and \code{NA} will be returned.}

\item{is_sorted}{Whether the data is pre-sorted in descending order. This may
speed up computations for some algorithms. The pre-sorted assumption is
tested and a warning is emitted if unsorted data is detected.}
}
\value{
The summary number.
}
\description{
Compute bibliometric indices such as the h-index, g-index, and i10-index.
}
\section{Implicit Type Conversions}{

The \code{h_index()} function implicitly coerces inputs into integer vectors,
which will truncate any floating point inputs. This usually will result in
expected outputs, as there are not typically fractional inputs in the
intended domain, and the definitions of these indices are defined on integral
thresholds explicitly. However, to maximize the versatility of g-index
computation, the \code{g_index()} function does not perform this integer coercion.
Therefore it is worth noting that floating point input can push the g-index
higher on edge cases. For example,
\code{g_index(as.integer(daily_totals$total_length)) != g_index(daily_totals$total_length)} Thus to ensure accurate g-index results
on data that may have a fractional component, it is advised to first perform
an integer conversion prior to passing a vector into \code{g_index()} or otherwise
validate inputs.

This integer conversion will also cause the \code{h_index()} to fail when inputs
contain extremely large values (\eqn{> 2^{31} - 1}{> 2^31 - 1}). The
Eddington number family of functions and \code{durfee()} do not have this check,
and may result in inaccurate outputs.
}

\references{
\url{https://en.wikipedia.org/wiki/Author-level_metrics},
\url{https://en.wikipedia.org/wiki/G-index}
}
\seealso{
\code{\link[=E_num]{E_num()}}, \code{\link[=durfee]{durfee()}}
}
